fix: Address critical hotkey issues from CodeRabbit review
- Fix keyCode=0 sentinel conflict that prevented binding "A" key
- Replace UInt16(0) with UInt16.max for modifier-only hotkey detection
- Eliminate code duplication by creating shared HotkeyUtilities
- Consolidate formatHotkeyDescription and keyCodeToDisplayString methods
- Reduce codebase by ~220 lines of duplicate code
- Ensure all hotkey formatting uses single source of truth

Fixes prevent "A" key binding issues and improve code maintainability.
